#75: Split off LayerDeltaUtil
